+++ cider-repl.el
@@ -1002,9 +1002,9 @@ constructs."
   (let ((map (make-sparse-keymap)))
     (set-keymap-parent map clojure-mode-map)
     (define-key map (kbd "C-c C-d") 'cider-doc-map)
-    (define-key map (kbd "M-.") 'cider-jump-to-var)
+    (define-key map (kbd "M-.") 'cider-find-var)
     (define-key map (kbd "M-,") 'cider-jump-back)
-    (define-key map (kbd "C-c M-.") 'cider-jump-to-resource)
+    (define-key map (kbd "C-c M-.") 'cider-find-resource)
     (define-key map (kbd "RET") 'cider-repl-return)
     (define-key map (kbd "TAB") 'cider-repl-tab)
     (define-key map (kbd "C-<return>") 'cider-repl-closing-return)
@@ -1045,8 +1045,8 @@ constructs."
         "--"
         ,cider-doc-menu
         "--"
-        ["Jump to source" cider-jump-to-var]
-        ["Jump to resource" cider-jump-to-resource]
+        ["Jump to source" cider-find-var]
+        ["Jump to resource" cider-find-resource]
         ["Jump back" cider-jump-back]
         ["Switch to Clojure buffer" cider-switch-to-last-clojure-buffer]
         "--"